backend/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
pillow
rembg
matplotlib
numpy